FIX[dmap-import]: Change pk_id to BigInteger for large tables. (#44)
This changes updates the pk_id column for large dmap tables to use BigInteger, with much larger maximum value.

+++ src/cubic_loader/dmap/schemas/sale_transaction.py
@@ -10,7 +10,7 @@ class SaleTransaction(SqlBase):
 
     __tablename__ = "sale_transaction"
 
-    pk_id = sa.Column(sa.Integer, primary_key=True)
+    pk_id = sa.Column(sa.BigInteger, primary_key=True)
     dataset_id = sa.Column(sa.String(), nullable=True, index=True)
     id = sa.Column(sa.String(), nullable=True)
     inserted_dtm = sa.Column(sa.DateTime, nullable=True)
